introduce checks in a separate module
