+++ test/specs/commonTests/isConformant.tsx
@@ -11,15 +11,7 @@ import helpers from './commonHelpers'
 import * as stardust from 'src/'
 import { felaRenderer } from 'src/lib'
 
-type ShorthandTestOptions = {
-  mapsValueToProp?: string
-}
-
-const DefaultShorthandTestOptions: ShorthandTestOptions = {
-  mapsValueToProp: 'content',
-}
-
-const mount = (node, options?) => {
+export const mount = (node, options?) => {
   return enzymeMount(
     <ThemeProvider theme={{ renderer: felaRenderer }}>{node}</ThemeProvider>,
     options,
@@ -507,47 +499,6 @@ export default (Component, options: any = {}) => {
       // test suites
       // -----------------------------------
       return this
-    },
-
-    hasConformantShorthandProperty(
-      shorthandPropertyName: string,
-      ShorthandComponent,
-      options: ShorthandTestOptions = DefaultShorthandTestOptions,
-    ) {
-      const { mapsValueToProp } = options
-
-      describe(`shorthand property for '${ShorthandComponent.displayName}'`, () => {
-        test(`is defined`, () => {
-          expect(Component.propTypes[shorthandPropertyName]).toBeTruthy()
-        })
-
-        test(`string value is handled as ${
-          ShorthandComponent.displayName
-        }'s ${mapsValueToProp}`, () => {
-          const props = { [shorthandPropertyName]: 'some value' }
-          const wrapper = mount(<Component {...props} />)
-
-          const shorthandComponentProps = wrapper.find(ShorthandComponent.displayName).props()
-          expect(shorthandComponentProps[mapsValueToProp]).toEqual('some value')
-        })
-
-        test(`object value is spread as ${ShorthandComponent.displayName}'s props`, () => {
-          const ShorthandValue = { foo: 'foo value', bar: 'bar value' }
-
-          const props = { [shorthandPropertyName]: ShorthandValue }
-          const wrapper = mount(<Component {...props} />)
-
-          const shorthandComponentProps = wrapper.find(ShorthandComponent.displayName).props()
-
-          const allShorthandPropertiesArePassedToShorthandComponent = Object.keys(
-            ShorthandValue,
-          ).every(
-            propertyName => ShorthandValue[propertyName] === shorthandComponentProps[propertyName],
-          )
-
-          expect(allShorthandPropertiesArePassedToShorthandComponent).toBe(true)
-        })
-      })
     },
   }
 }
