+++ test/specs/components/Button/Button-test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { isConformant, handlesAccessibility } from 'test/specs/commonTests'
+import { isConformant, handlesAccessibility, implementsShorthandProp } from 'test/specs/commonTests'
 import { getTestingRenderedComponent, mountWithProvider } from 'test/utils'
 
 import Button from 'src/components/Button/Button'
@@ -8,10 +8,11 @@ import Icon from 'src/components/Icon/Icon'
 
 import { MenuBehavior } from 'src/lib/accessibility'
 
+const buttonImplementsShorthandProp = implementsShorthandProp(Button)
+
 describe('Button', () => {
-  isConformant(Button).hasConformantShorthandProperty('icon', Icon, {
-    mapsValueToProp: 'name',
-  })
+  isConformant(Button)
+  buttonImplementsShorthandProp('icon', Icon, { mapsValueToProp: 'name' })
 
   handlesAccessibility(Button, {
     defaultRootRole: 'button',
